pyproject.toml
[project]
name = "bilingual-arabic-english-rag-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "datasets>=4.5.0",
    "easyocr>=1.7.2",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.8",
    "langgraph-prebuilt>=1.0.7",
    "openai>=2.21.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.3",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "wikipedia>=1.4.0",
]
